clean up audio data a bit
