+++ audio/instruments.c
@@ -1,8 +1,6 @@
 #include "global.h"
 #include "sound.h"
 
-#include "instruments.inc.c"
+#include "instrument_data.inc.c"
 #include "instrument_banks.inc.c"
 #include "instrument_bank_list.inc.c"
-#include "sequence_names.inc.c"
-#include "sequence_data.inc.c"
