+++ include/sound.h
@@ -53,6 +53,120 @@ enum DirectSoundModesEnum {
     DIRECTSOUND_MODE_MONO2
 };
 
+enum InstrumentBanksEnum {
+	INST_BANK_UNUSED_0,
+	INST_BANK_UNUSED_1,
+	INST_BANK_UNUSED_2,
+	INST_BANK_UNUSED_3,
+	INST_BANK_UNUSED_4,
+	INST_BANK_54,
+	INST_BANK_UNUSED_5,
+	INST_BANK_UNUSED_7,
+	INST_BANK_UNUSED_8,
+	INST_BANK_UNUSED_9,
+	INST_BANK_UNUSED_10,
+	INST_BANK_UNUSED_11,
+	INST_BANK_UNUSED_12,
+	INST_BANK_UNUSED_13,
+	INST_BANK_UNUSED_14,
+	INST_BANK_UNUSED_15,
+	INST_BANK_UNUSED_16,
+	INST_BANK_55,
+	INST_BANK_1,
+	INST_BANK_2,
+	INST_BANK_3,
+	INST_BANK_4,
+	INST_BANK_5,
+	INST_BANK_6,
+	INST_BANK_7,
+	INST_BANK_8,
+	INST_BANK_9,
+	INST_BANK_10,
+	INST_BANK_11,
+	INST_BANK_12,
+	INST_BANK_13,
+	INST_BANK_14,
+	INST_BANK_15,
+	INST_BANK_16,
+	INST_BANK_17,
+	INST_BANK_UNUSED_35,
+	INST_BANK_UNUSED_36,
+	INST_BANK_UNUSED_37,
+	INST_BANK_UNUSED_38,
+	INST_BANK_UNUSED_39,
+	INST_BANK_47,
+	INST_BANK_49,
+	INST_BANK_50,
+	INST_BANK_48,
+	INST_BANK_UNUSED_44,
+	INST_BANK_51,
+	INST_BANK_52,
+	INST_BANK_53,
+	INST_BANK_UNUSED_48,
+	INST_BANK_UNUSED_49,
+	INST_BANK_18,
+	INST_BANK_19,
+	INST_BANK_20,
+	INST_BANK_21,
+	INST_BANK_22,
+	INST_BANK_23,
+	INST_BANK_24,
+	INST_BANK_25,
+	INST_BANK_26,
+	INST_BANK_27,
+	INST_BANK_28,
+	INST_BANK_29,
+	INST_BANK_30,
+	INST_BANK_31,
+	INST_BANK_32,
+	INST_BANK_33,
+	INST_BANK_34,
+	INST_BANK_35,
+	INST_BANK_36,
+	INST_BANK_37,
+	INST_BANK_38,
+	INST_BANK_39,
+	INST_BANK_40,
+	INST_BANK_41,
+	INST_BANK_42,
+	INST_BANK_43,
+	INST_BANK_44,
+	INST_BANK_45,
+	INST_BANK_46,
+	INST_BANK_UNUSED_79,
+	INST_BANK_UNUSED_80,
+	INST_BANK_UNUSED_81,
+	INST_BANK_UNUSED_82,
+	INST_BANK_UNUSED_83,
+	INST_BANK_UNUSED_84,
+	INST_BANK_UNUSED_85,
+	INST_BANK_UNUSED_86,
+	INST_BANK_UNUSED_87,
+	INST_BANK_UNUSED_88,
+	INST_BANK_UNUSED_89,
+	INST_BANK_UNUSED_90,
+	INST_BANK_UNUSED_91,
+	INST_BANK_UNUSED_92,
+	INST_BANK_UNUSED_93,
+	INST_BANK_UNUSED_94,
+	INST_BANK_UNUSED_95,
+	INST_BANK_UNUSED_96,
+	INST_BANK_UNUSED_97,
+	INST_BANK_UNUSED_98,
+	INST_BANK_UNUSED_99,
+	INST_BANK_56,
+	INST_BANK_57,
+	INST_BANK_58,
+	INST_BANK_59,
+	INST_BANK_60,
+	INST_BANK_61,
+	INST_BANK_UNUSED_106,
+	INST_BANK_UNUSED_107,
+	INST_BANK_UNUSED_108,
+	INST_BANK_UNUSED_109,
+	INST_BANK_62,
+	INST_BANK_63
+};
 
 struct SampleInfo {
 	u32 length;
